feat: integrate marketing component into landing page
Delete '(marketing)' folder, move 'HowItWorks' to MDX components, update 'home.mdx'.
