+++ apps/web/src/content/pages/home.mdx
@@ -32,6 +32,8 @@ faq:
 
 <Image src="/assets/landing/dashboard.png" alt="dashboard monitor overview" priority />
 
+<HowItWorks />
+
 ## Trusted by teams who value transparency
 
 <Grid cols={4}>
